--- env.json
@@ -1,3 +1,0 @@
-{
-  "mode": "production"
-}
