+++ src/js/main.js
@@ -1,9 +1,3 @@
-const env = require('../../env.json')
-
-if (env.mode) {
-  //process.env.NODE_ENV = env.mode
-}
-
 const {app, ipcMain, BrowserWindow, dialog, powerSaveBlocker} = electron = require('electron')
 
 const fs = require('fs-extra')
@@ -36,7 +30,7 @@ const MobileServer = require('./express-app/app')
 
 const preferencesUI = require('./windows/preferences')()
 const registration = require('./windows/registration/main')
-const shotGeneratorWindow = require('./windows/shot-generator-new/main')
+const shotGeneratorWindow = require('./windows/shot-generator/main')
 const tutorialMain = require('./windows/shot-generator-tutorial/main')
 
 const JWT = require('jsonwebtoken')
