+++ src/js/shot-generator/Image.js
@@ -11,7 +11,7 @@ const textureLoader = new THREE.TextureLoader()
 const IconSprites = require('./IconSprites')
 
 const { isUserFile } = require('../services/model-loader')
-const pathToShotGeneratorData = path.join(__dirname, '..', '..', '..', 'src', 'data', 'shot-generator')
+const pathToShotGeneratorData = path.join(window.__dirname, '..', 'src', 'data', 'shot-generator')
 const pathToBuiltInVolumeImages = path.join(pathToShotGeneratorData, 'images')
 
 const materialFactory = () => new THREE.MeshToonMaterial({
